MAT-9448 removed unwanted code and fixed the broken stratifications
